+++ ai-services/internal/pkg/catalog/client/application.go
@@ -28,18 +28,6 @@ func NewApplicationClient() (*ApplicationClient, error) {
 	}, nil
 }
 
-// ListApplicationsParams holds optional query parameters for listing applications.
-type ListApplicationsParams struct {
-	// Page is the page number (1-indexed). Default: 1
-	Page int
-	// PageSize is the number of items per page (max: 100). Default: 20
-	PageSize int
-	// DeploymentType filters by deployment type: 'architectures' or 'services'
-	DeploymentType string
-	// CatalogID filters by catalog ID (e.g., 'rag', 'chat', 'digitize', 'summarize')
-	CatalogID string
-}
-
 // ListApplications retrieves a paginated list of all applications for the authenticated user.
 // It supports optional filters via the params argument.
 //
