+++ ai-services/internal/pkg/catalog/client/models.go
@@ -1,0 +1,13 @@
+package client
+
+// ListApplicationsParams holds optional query parameters for listing applications.
+type ListApplicationsParams struct {
+	// Page is the page number (1-indexed). Default: 1
+	Page int
+	// PageSize is the number of items per page (max: 100). Default: 20
+	PageSize int
+	// DeploymentType filters by deployment type: 'architectures' or 'services'
+	DeploymentType string
+	// CatalogID filters by catalog ID (e.g., 'rag', 'chat', 'digitize', 'summarize')
+	CatalogID string
+}
